pre-mhd python test refactor
