+++ tests/simulator/advance/test_fields_advance_2d.py
@@ -4,56 +4,70 @@ This file exists independently from test_advance.py to isolate dimension
 """
 
 import unittest
-
-import matplotlib
 from ddt import data, ddt, unpack
+
+import pyphare.pharein as ph
 from pyphare.core.box import Box2D
 
-from tests.simulator.test_advance import AdvanceTestBase
+from tests.simulator.advance.test_advance_mhd import MHDAdvanceTest
+from tests.simulator.advance.test_advance_hybrid import HybridAdvanceTest
 
-matplotlib.use("Agg")  # for systems without GUI
-
+ph.NO_GUI()
 ndim = 2
 interp_orders = [1, 2, 3]
 ppc = 25
 
 
-def per_interp(dic):
-    return [(interp, dic) for interp in interp_orders]
+def permute_hybrid(boxes={}):
+    return [
+        dict(
+            super_class=HybridAdvanceTest,
+            interp_order=interp_order,
+            refinement_boxes=boxes,
+            nbr_part_per_cell=ppc,
+        )
+        for interp_order in interp_orders
+    ]
+
+
+def permute_mhd(boxes={}):  # interp_order hax todo
+    return [dict(super_class=MHDAdvanceTest, interp_order=2, refinement_boxes=boxes)]
+
+
+def permute(boxes={}, hybrid=True, mhd=False):
+    return (permute_hybrid(boxes) if hybrid else []) + (
+        permute_mhd(boxes) if mhd else []
+    )
 
 
 @ddt
-class AdvanceTest(AdvanceTestBase):
+class AdvanceTest2D(HybridAdvanceTest, MHDAdvanceTest):
     @data(
-        *per_interp({}),
-        *per_interp({"L0": [Box2D(10, 19)]}),
-        *per_interp({"L0": [Box2D(8, 20)]}),
+        *permute({}),
+        *permute({"L0": [Box2D(10, 19)]}),
+        *permute({"L0": [Box2D(8, 20)]}),
     )
     @unpack
-    def test_overlaped_fields_are_equal(self, interp_order, refinement_boxes):
+    def test_overlaped_fields_are_equal(self, super_class, **kwargs):
+        self.__class__ = super_class  # cast to super class
         print(f"{self._testMethodName}_{ndim}d")
         time_step_nbr = 3
         time_step = 0.001
 
         datahier = self.getHierarchy(
-            ndim,
-            interp_order,
-            refinement_boxes,
-            "eb",
-            time_step=time_step,
-            time_step_nbr=time_step_nbr,
-            nbr_part_per_cell=ppc,
+            ndim, qty="eb", time_step=time_step, time_step_nbr=time_step_nbr, **kwargs
         )
         self._test_overlaped_fields_are_equal(datahier, time_step_nbr, time_step)
 
     @data(
-        *per_interp({}),
-        *per_interp({"L0": [Box2D(10, 19)]}),
+        *permute({}),
+        *permute({"L0": [Box2D(10, 19)]}),
     )
     @unpack
     def test_overlaped_fields_are_equal_with_min_max_patch_size_of_max_ghosts(
-        self, interp_order, refinement_boxes
+        self, super_class, interp_order, **kwargs
     ):
+        self.__class__ = super_class  # cast to super class
         print(f"{self._testMethodName}_{ndim}d")
         time_step_nbr = 3
         time_step = 0.001
@@ -65,23 +79,22 @@ class AdvanceTest(AdvanceTestBase):
         datahier = self.getHierarchy(
             ndim,
             interp_order,
-            refinement_boxes,
-            "eb",
+            qty="eb",
             smallest_patch_size=smallest_patch_size,
             largest_patch_size=smallest_patch_size,
             time_step=time_step,
             time_step_nbr=time_step_nbr,
-            nbr_part_per_cell=ppc,
+            **kwargs,
         )
         self._test_overlaped_fields_are_equal(datahier, time_step_nbr, time_step)
 
     @data(
-        *per_interp(({"L0": {"B0": Box2D(10, 14)}})),
-        *per_interp(({"L0": {"B0": Box2D(10, 14), "B1": Box2D(15, 19)}})),
-        *per_interp(({"L0": {"B0": Box2D(6, 23)}})),
-        *per_interp(({"L0": {"B0": Box2D(2, 12), "B1": Box2D(13, 25)}})),
-        *per_interp(({"L0": {"B0": Box2D(5, 20)}, "L1": {"B0": Box2D(15, 19)}})),
-        *per_interp(
+        *permute(({"L0": {"B0": Box2D(10, 14)}})),
+        *permute(({"L0": {"B0": Box2D(10, 14), "B1": Box2D(15, 19)}})),
+        *permute(({"L0": {"B0": Box2D(6, 23)}})),
+        *permute(({"L0": {"B0": Box2D(2, 12), "B1": Box2D(13, 25)}})),
+        *permute(({"L0": {"B0": Box2D(5, 20)}, "L1": {"B0": Box2D(15, 19)}})),
+        *permute(
             (
                 {
                     "L0": {"B0": Box2D(5, 20)},
@@ -92,27 +105,27 @@ class AdvanceTest(AdvanceTestBase):
         ),
     )
     @unpack
-    def test_field_coarsening_via_subcycles(self, interp_order, refinement_boxes):
+    def test_field_coarsening_via_subcycles(self, super_class, **kwargs):
+        self.__class__ = super_class  # cast to super class
         print(f"{self._testMethodName}_{ndim}d")
-        self._test_field_coarsening_via_subcycles(
-            ndim, interp_order, refinement_boxes, dl=0.3
-        )
+        self._test_field_coarsening_via_subcycles(ndim, dl=0.3, **kwargs)
 
     @unittest.skip("should change to work with moments")
     @data(  # only supports a hierarchy with 2 levels
-        *per_interp(({"L0": [Box2D(0, 4)]})),
-        *per_interp(({"L0": [Box2D(10, 14)]})),
-        *per_interp(({"L0": [Box2D(0, 4), Box2D(10, 14)]})),
-        *per_interp(({"L0": [Box2D(0, 4), Box2D(5, 9), Box2D(10, 14)]})),
-        *per_interp(({"L0": [Box2D(20, 24)]})),
+        *permute(({"L0": [Box2D(0, 4)]})),
+        *permute(({"L0": [Box2D(10, 14)]})),
+        *permute(({"L0": [Box2D(0, 4), Box2D(10, 14)]})),
+        *permute(({"L0": [Box2D(0, 4), Box2D(5, 9), Box2D(10, 14)]})),
+        *permute(({"L0": [Box2D(20, 24)]})),
     )
     @unpack
     def test_field_level_ghosts_via_subcycles_and_coarser_interpolation(
-        self, interp_order, refinement_boxes
+        self, super_class, **kwargs
     ):
+        self.__class__ = super_class  # cast to super class
         print(f"{self._testMethodName}_{ndim}d")
         self._test_field_level_ghosts_via_subcycles_and_coarser_interpolation(
-            ndim, interp_order, refinement_boxes
+            ndim, **kwargs
         )
