+++ tests/simulator/initialize/test_particles_init_1d.py
@@ -1,17 +1,17 @@
 """
-  This file exists independently from test_initialization.py to isolate dimension
-    test cases and allow each to be overridden in some way if required.
+This file exists independently from test_initialization.py to isolate dimension
+  test cases and allow each to be overridden in some way if required.
 """
 
 import unittest
-
-import matplotlib
 from ddt import data, ddt, unpack
+
+import pyphare.pharein as ph
 from pyphare.core.box import Box1D
 
-from tests.simulator.test_initialization import InitializationTest
+from tests.simulator.initialize.test_init_hybrid import HybridInitializationTest
 
-matplotlib.use("Agg")  # for systems without GUI
+ph.NO_GUI()
 
 ndim = 1
 interp_orders = [1, 2, 3]
@@ -23,7 +23,7 @@ def per_interp(dic):
 
 
 @ddt
-class Initialization1DTest(InitializationTest):
+class Initialization1DTest(HybridInitializationTest):
     @data(*interp_orders)
     def test_nbr_particles_per_cell_is_as_provided(self, interp_order):
         print(f"{self._testMethodName}_{ndim}d")
