+++ tests/simulator/refinement/test_2d_10_core.py
@@ -1,13 +1,12 @@
 #!/usr/bin/env python3
 
 """
-  Complex test to force Yee Lattice centering discrepancies due to SAMRAI Schedule/Algorithm
-    grouping of quantity components with distinct geometries.
+Complex test to force Yee Lattice centering discrepancies due to SAMRAI Schedule/Algorithm
+  grouping of quantity components with distinct geometries.
 
-  Should be run with 10 cores
-    mpirun -n 10 tests/simulator/refinement/test_2d_10_core.py
+Should be run with 10 cores
+  mpirun -n 10 tests/simulator/refinement/test_2d_10_core.py
 """
-
 
 import numpy as np
