+++ tests/simulator/test_advance.py
@@ -1,5 +1,10 @@
 #
+#  Common base class across hybrid and mhd tests
+#   see
+#      tests/simulator/advance/test_advance_mhd.py
+#      tests/simulator/advance/test_advance_hybrid.py
 #
+
 
 import unittest
 import numpy as np
@@ -12,253 +17,15 @@ from pyphare.core.box import amr_to_local
 from pyphare import cpp
 import pyphare.core.box as boxm
 from pyphare.core.box import Box
-from pyphare.core.box import amr_to_local
-from pyphare.core.phare_utilities import assert_fp_any_all_close, np_array_ify
-from pyphare.pharein import ElectronModel, MaxwellianFluidModel
-from pyphare.pharein.diagnostics import (
-    ElectromagDiagnostics,
-    FluidDiagnostics,
-    ParticleDiagnostics,
-)
-from pyphare.pharein.simulation import Simulation
+from pyphare.core.phare_utilities import assert_fp_any_all_close
 from pyphare.pharesee.geometry import hierarchy_overlaps, level_ghost_boxes
-from pyphare.pharesee.hierarchy import hierarchy_from
 from pyphare.pharesee.hierarchy.hierarchy import format_timestamp
-from pyphare.pharesee.hierarchy.hierarchy_utils import merge_particles
-from pyphare.simulator.simulator import Simulator
 
-from tests.diagnostic import all_timestamps
 from tests.simulator import SimulatorTest, diff_boxes
 
 
 @ddt
 class AdvanceTestBase(SimulatorTest):
-    """
-    This class groups the setup of tests and the implementation of tests that is
-    common regardless of the dimensionality.
-
-    dimension dependent aspects are to be found in:
-
-    - advance/test_field_advance_1d.py
-    - advance/test_field_advance_2d.py
-    - advance/test_field_advance_3d.py
-
-    """
-
-    # ----------------------------------------------------------------------
-    #
-    #
-    #                       TEST SETUP
-    #
-    #
-    # ----------------------------------------------------------------------
-
-    def _density(*xyz):
-        from pyphare.pharein.global_vars import sim
-
-        hL = np.array(sim.simulation_domain()) / 2
-        _ = lambda i: -((xyz[i] - hL[i]) ** 2)
-        return 0.3 + np.exp(sum([_(i) for i in range(len(xyz))]))
-
-    def getHierarchy(
-        self,
-        ndim,
-        interp_order,
-        refinement_boxes,
-        qty,
-        nbr_part_per_cell=100,
-        density=_density,
-        smallest_patch_size=None,
-        largest_patch_size=20,
-        cells=120,
-        time_step=0.001,
-        model_init={},
-        dl=0.2,
-        extra_diag_options={},
-        time_step_nbr=1,
-        timestamps=None,
-        block_merging_particles=False,
-        diag_outputs="",
-    ):
-        """
-        this function creates and run a simulation setup for the tests
-        and returns the hierarchy with all quantities requested.
-        """
-
-        diag_outputs = self.unique_diag_dir_for_test_case(
-            "phare_outputs/advance", ndim, interp_order, diag_outputs
-        )
-
-        from pyphare.pharein import global_vars
-
-        global_vars.sim = None
-        if smallest_patch_size is None:
-            from pyphare.pharein.simulation import check_patch_size
-
-            _, smallest_patch_size = check_patch_size(
-                ndim, interp_order=interp_order, cells=cells
-            )
-
-        # ----------------------------------------------------------------------
-        # simulation setup and running
-        # ----------------------------------------------------------------------
-
-        extra_diag_options["mode"] = "overwrite"
-        extra_diag_options["dir"] = diag_outputs
-
-        self.register_diag_dir_for_cleanup(diag_outputs)
-        sim = Simulation(
-            smallest_patch_size=smallest_patch_size,
-            largest_patch_size=largest_patch_size,
-            time_step_nbr=time_step_nbr,
-            time_step=time_step,
-            boundary_types=["periodic"] * ndim,
-            cells=np_array_ify(cells, ndim),
-            dl=np_array_ify(dl, ndim),
-            interp_order=interp_order,
-            refinement_boxes=refinement_boxes,
-            diag_options={"format": "phareh5", "options": extra_diag_options},
-            strict=True,
-        )
-
-        def bx(*xyz):
-            return 1.0
-
-        def by(*xyz):
-            L = sim.simulation_domain()
-            _ = lambda i: 0.1 * np.cos(2 * np.pi * xyz[i] / L[i])
-            return np.asarray([_(i) for i in range(len(xyz))]).prod(axis=0)
-
-        def bz(*xyz):
-            L = sim.simulation_domain()
-            _ = lambda i: 0.1 * np.cos(2 * np.pi * xyz[i] / L[i])
-            return np.asarray([_(i) for i in range(len(xyz))]).prod(axis=0)
-
-        def vx(*xyz):
-            L = sim.simulation_domain()
-            _ = lambda i: 0.1 * np.cos(2 * np.pi * xyz[i] / L[i])
-            return np.asarray([_(i) for i in range(len(xyz))]).prod(axis=0)
-
-        def vy(*xyz):
-            L = sim.simulation_domain()
-            _ = lambda i: 0.1 * np.cos(2 * np.pi * xyz[i] / L[i])
-            return np.asarray([_(i) for i in range(len(xyz))]).prod(axis=0)
-
-        def vz(*xyz):
-            L = sim.simulation_domain()
-            _ = lambda i: 0.1 * np.cos(2 * np.pi * xyz[i] / L[i])
-            return np.asarray([_(i) for i in range(len(xyz))]).prod(axis=0)
-
-        def vth(*xyz):
-            return 0.01 + np.zeros_like(xyz[0])
-
-        def vthx(*xyz):
-            return vth(*xyz)
-
-        def vthy(*xyz):
-            return vth(*xyz)
-
-        def vthz(*xyz):
-            return vth(*xyz)
-
-        MaxwellianFluidModel(
-            bx=bx,
-            by=by,
-            bz=bz,
-            protons={
-                "charge": 1,
-                "density": density,
-                "vbulkx": vx,
-                "vbulky": vy,
-                "vbulkz": vz,
-                "vthx": vthx,
-                "vthy": vthy,
-                "vthz": vthz,
-                "nbr_part_per_cell": nbr_part_per_cell,
-                "init": model_init,
-            },
-        )
-
-        ElectronModel(closure="isothermal", Te=0.12)
-
-        if timestamps is None:
-            timestamps = all_timestamps(global_vars.sim)
-
-        for quantity in ["E", "B"]:
-            ElectromagDiagnostics(quantity=quantity, write_timestamps=timestamps)
-
-        for quantity in ["charge_density", "bulkVelocity"]:
-            FluidDiagnostics(
-                quantity=quantity,
-                write_timestamps=timestamps,
-            )
-
-        poplist = ["protons"]
-        for pop in poplist:
-            for quantity in ["density", "flux"]:
-                FluidDiagnostics(
-                    quantity=quantity,
-                    write_timestamps=timestamps,
-                    population_name=pop,
-                )
-
-            for quantity in ["domain", "levelGhost"]:
-                ParticleDiagnostics(
-                    quantity=quantity, write_timestamps=timestamps, population_name=pop
-                )
-
-        Simulator(global_vars.sim).run()
-
-        # ----------------------------------------------------------------------
-        # The simulation has run, now we build the hierarchy with the requested
-        # quantities.
-        # ----------------------------------------------------------------------
-
-        eb_hier = None
-        if qty in ["e", "eb", "fields"]:
-            eb_hier = hierarchy_from(
-                h5_filename=diag_outputs + "/EM_E.h5", hier=eb_hier
-            )
-        if qty in ["b", "eb", "fields"]:
-            eb_hier = hierarchy_from(
-                h5_filename=diag_outputs + "/EM_B.h5", hier=eb_hier
-            )
-        if qty in ["e", "b", "eb"]:
-            return eb_hier
-
-        if qty == "moments" or qty == "fields":
-            mom_hier = hierarchy_from(
-                h5_filename=diag_outputs + "/ions_charge_density.h5", hier=eb_hier
-            )
-            mom_hier = hierarchy_from(
-                h5_filename=diag_outputs + "/ions_bulkVelocity.h5", hier=mom_hier
-            )
-            return mom_hier
-
-        # else particle tests
-
-        assert qty == "particles"
-
-        particle_hier = hierarchy_from(
-            h5_filename=diag_outputs + "/ions_pop_protons_domain.h5"
-        )
-        particle_hier = hierarchy_from(
-            h5_filename=diag_outputs + "/ions_pop_protons_levelGhost.h5",
-            hier=particle_hier,
-        )
-
-        if not block_merging_particles:
-            merge_particles(particle_hier)
-        return particle_hier
-
-        # ----------------------------------------------------------------------
-        #
-        #
-        #                       TEST DEFINITIONS
-        #
-        #
-        # ----------------------------------------------------------------------
-
     def base_test_overlaped_fields_are_equal(self, datahier, coarsest_time):
         """
         here overlaps are calculated between patches at the same level
@@ -310,6 +77,7 @@ class AdvanceTestBase(SimulatorTest):
 
                     if ovrlp_box.ndim == 1:
                         failed_i = np.where(np.abs(slice1 - slice2) > 5.5e-15)
+
 
                     if ovrlp_box.ndim == 2:
                         failed_i, failed_j = np.where(np.abs(slice1 - slice2) > 5.5e-15)
@@ -432,96 +200,7 @@ class AdvanceTestBase(SimulatorTest):
         self.assertGreater(successful_test_nbr, time_step_nbr)
         self.assertEqual(successful_test_nbr % (time_step_nbr + 1), 0)
 
-    def _test_overlapped_particledatas_have_identical_particles(
-        self, ndim, interp_order, refinement_boxes, ppc=100, **kwargs
-    ):
-        print(
-            "test_overlapped_particledatas_have_identical_particles, interporder : {}".format(
-                interp_order
-            )
-        )
-        from copy import copy
-
-        time_step_nbr = 3
-        time_step = 0.001
-
-        datahier = self.getHierarchy(
-            ndim,
-            interp_order,
-            refinement_boxes,
-            "particles",
-            time_step=time_step,
-            time_step_nbr=time_step_nbr,
-            nbr_part_per_cell=ppc,
-            **kwargs,
-        )
-
-        for time_step_idx in range(time_step_nbr + 1):
-            coarsest_time = time_step_idx * time_step
-
-            overlaps = hierarchy_overlaps(datahier, coarsest_time)
-
-            for ilvl in datahier.levels():
-                print("testing level {}".format(ilvl))
-                for overlap in overlaps[ilvl]:
-                    pd1, pd2 = overlap["pdatas"]
-                    box = overlap["box"]
-                    offsets = overlap["offset"]
-
-                    self.assertEqual(pd1.quantity, pd2.quantity)
-
-                    if "particles" in pd1.quantity:
-                        # the following uses 'offset', we need to remember that offset
-                        # is the quantity by which a patch has been moved to detect
-                        # overlap with the other one.
-                        # so shift by +offset when evaluating patch data in overlap box
-                        # index space, and by -offset when we want to shift box indexes
-                        # to the associated patch index space.
-
-                        # overlap box must be shifted by -offset to select data in the patches
-                        part1 = copy(
-                            pd1.dataset.select(boxm.shift(box, -np.asarray(offsets[0])))
-                        )
-                        part2 = copy(
-                            pd2.dataset.select(boxm.shift(box, -np.asarray(offsets[1])))
-                        )
-
-                        # periodic icell overlaps need shifting to be the same
-                        part1.iCells = part1.iCells + offsets[0]
-                        part2.iCells = part2.iCells + offsets[1]
-                        self.assertEqual(part1, part2)
-
-    def _test_L0_particle_number_conservation(
-        self, ndim, interp_order, ppc=100, cells=120
-    ):
-        time_step_nbr = 10
-        time_step = 0.001
-
-        n_particles = ppc * (cells**ndim)
-
-        datahier = self.getHierarchy(
-            ndim,
-            interp_order,
-            None,
-            "particles",
-            time_step=time_step,
-            time_step_nbr=time_step_nbr,
-            nbr_part_per_cell=ppc,
-            cells=cells,
-        )
-
-        for time_step_idx in range(time_step_nbr + 1):
-            coarsest_time = time_step_idx * time_step
-            n_particles_at_t = 0
-            for patch in datahier.level(0, coarsest_time).patches:
-                n_particles_at_t += (
-                    patch.patch_datas["protons_particles"].dataset[patch.box].size()
-                )
-            self.assertEqual(n_particles, n_particles_at_t)
-
-    def _test_field_coarsening_via_subcycles(
-        self, dim, interp_order, refinement_boxes, cells=60, **kwargs
-    ):
+    def _test_field_coarsening_via_subcycles(self, dim, interp_order, **kwargs):
         print(
             "test_field_coarsening_via_subcycles for dim/interp : {}/{}".format(
                 dim, interp_order
@@ -534,14 +213,11 @@ class AdvanceTestBase(SimulatorTest):
 
         time_step_nbr = 3
 
-        diag_outputs = f"subcycle_coarsening/{dim}/{interp_order}/{self.ddt_test_id()}"
         datahier = self.getHierarchy(
             dim,
             interp_order,
-            refinement_boxes,
-            "fields",
-            cells=cells,
-            diag_outputs=diag_outputs,
+            qty="fields",
+            cells=60,
             time_step=0.001,
             extra_diag_options={"fine_dump_lvl_max": 10},
             time_step_nbr=time_step_nbr,
@@ -888,39 +564,6 @@ class AdvanceTestBase(SimulatorTest):
             successful_test_nbr, len(refinement_boxes["L0"]) * len(quantities)
         )
 
-    def base_test_domain_particles_on_refined_level(self, datahier, new_time=None):
-        """
-        !! test assumes only domain particle patch_datas are present !!
-        """
-        times = datahier.times() if new_time is None else [new_time]
-        successful_test_nbr = 0
-        for coarsest_time in times:
-            for patch in datahier.level(1, coarsest_time).patches:
-                for pd_key, pd in patch.patch_datas.items():
-                    if pd_key.endswith("_domain"):
-                        self.assertGreater(pd[pd.box].size(), 0)
-                        self.assertEqual(pd[pd.box].size(), pd[pd.ghost_box].size())
-                        successful_test_nbr += 1
-        self.assertGreater(successful_test_nbr, 0)
-
-    def _test_domain_particles_on_refined_level(
-        self, ndim, interp_order, refinement_boxes, **kwargs
-    ):
-        time_step_nbr = 5
-        time_step = 0.001
-
-        self.base_test_domain_particles_on_refined_level(
-            self.getHierarchy(
-                ndim,
-                interp_order,
-                refinement_boxes,
-                qty="particles",
-                time_step=time_step,
-                time_step_nbr=time_step_nbr,
-                block_merging_particles=True,
-                **kwargs,
-            )
-        )
 
 
 if __name__ == "__main__":
